- fixed editor connection

+++ runtime/cpp/putki/liveupdate/liveupdate.cpp
@@ -62,7 +62,7 @@ namespace putki
 			bool connected;
 
 			char readbuf[READBUF_SIZE];
-			unsigned long readpos;
+			size_t readpos;
 
 			struct pkg_e
 			{
@@ -136,7 +136,7 @@ namespace putki
 		{
 			std::string cmd(str);
 			cmd.append("\n");
-			send(d->socket, cmd.c_str(), cmd.size(), 0);
+			send(d->socket, cmd.c_str(), (int)cmd.size(), 0);
 		}
 
 		data* connect()
@@ -147,7 +147,7 @@ namespace putki
 			addrLocal.sin_family = AF_INET;
 			addrLocal.sin_port = htons(5556);
 			addrLocal.sin_addr.s_addr = htonl(0x7f000001);
-			d->socket = socket(AF_INET, SOCK_STREAM, 0);
+			d->socket = (int)socket(AF_INET, SOCK_STREAM, 0);
 			if (connect(d->socket, (sockaddr*)&addrLocal, sizeof(addrLocal)) < 0)
 			{
 				PTK_WARNING("Could not connect to live update socket")
@@ -323,7 +323,7 @@ namespace putki
 				return;
 			}
 			
-			uint32_t hdr_size, data_size;
+			size_t hdr_size, data_size;
 			if (!pkgmgr::get_header_info(d->readbuf, d->readbuf + d->readpos, &hdr_size, &data_size))
 				return;
 				
@@ -379,8 +379,8 @@ namespace putki
 			}
 
 			// peel off this package.
-			unsigned long peel = hdr_size + data_size;
-			for (unsigned long bk=0; bk<(d->readpos-peel); bk++)
+			size_t peel = hdr_size + data_size;
+			for (size_t bk = 0; bk < (d->readpos - peel); bk++)
 				d->readbuf[bk] = d->readbuf[bk + peel];
 			d->readpos -= peel;
 
@@ -407,7 +407,7 @@ namespace putki
 			while (d->connected && select(d->socket + 1, &fds, (fd_set *) 0, (fd_set *) 0, &tv))
 			{
 				int read;
-				if ((read = recv(d->socket, &d->readbuf[d->readpos], READBUF_SIZE - d->readpos, 0)) > 0)
+				if ((read = recv(d->socket, &d->readbuf[d->readpos], (int)(READBUF_SIZE - d->readpos), 0)) > 0)
 				{
 					d->readpos += read;
 					on_recv(d);
